add preStop hook to sidecar container
Inject a preStop hook into the sidecar container that invokes the
kar/kill route to enable an orderly shutdown. Also define the
KAR_POD_IP envvar in the sidecar container using the downward API to
make the IP address of the pod available to the sidecar (not currently
used, but perhaps useful in the future).