suppress stderr output from embedded kubectl commands
We have to use kubectl get to check for the presence/absence
of resources.  If the resource is absence, kubectl prints a
message on stderr that can be confusing to users of the kar
scripts.  Redirect all kubectl stderr output to /dev/null to
get control over the script output.

Fixes #43.