halt current reminder processing round if TellActor returns an error (#63)
If TellActor fails, then put the reminder back in the in-memory queue
and end the current round of processing.  This ensures that the a
future attempt will be made to deliver the reminder (either in this
sidecar or someother one if the error from TellActor
is because the sidecar is shutting down).