standardize actor call convention for method parameters
The request body for an actor call must be a JSON Array containing
the arguments to use when invoking the target actor method. An empty
array is allowed (for methods that take no arguments). Currently
JavaScript supports an arbitrary number of parameters to actor methods
while Java is limited to 0 or 1 parameter.