bug fix in actor scheduling for recovery from a blocked self-call
During recovery, when an message is identified as needing to be
blocked to wait for a child call to complete it is necessary to
eagerly set the activeFlow on the freshly created actor instance. This
recreates the locking state before recovery and allows re-entrant
calls to be properly identified as they are dequeued from kafka.

Also extend fault-tolerance-actor test case to include self-calls.